Add files via upload
This exercise is about using control flow statements in JavaScript to make decisions based on conditions. This involves using if, else if, and else statements to execute different blocks of code depending on whether certain conditions are true or false. These control flow statements allow for dynamic decision-making in programs, such as determining race numbers and start times for runners based on their age and registration status.